Consolidate ADR-0039 dead-letter machinery onto StreamConsumer
Move the duplicated bounded-delivery reclaim/dead-letter machinery
(_reclaim_once, _dead_letter_over_cap, _dead_letter, the entry-fields
fetch, _ack) out of both Consumer and EvalStreamConsumer onto their
shared StreamConsumer base, parameterized per lane by a DeliverySpec
(stream/group/consumer, dead-letter target, over-cap reason, handler,
and per-lane logger + log-format strings). Each subclass keeps only its
handler and its own read/maintenance loops.

Unify the drifted over-cap reason string on "max-delivery-exceeded"
(the runs lane previously wrote "max delivery exceeded"). The runs
dead-letter log is re-emitted as a logfmt key=value critical alert, so a
space-separated reason value corrupts field parsing; the kebab token is
logfmt-safe and matches both the sibling "unparseable" reason and the
eval lane, which already used it.

Behavior is byte-identical per lane otherwise: XPENDING-before-XAUTOCLAIM,
cap at >=, the IDLE filter equal to min_idle_time, the _inflight_ids skip
ahead of the cap check, XADD-before-XACK, the PEL-not-memory delivery
count, whole-list paging, and the bounded graveyard are all preserved.
The runs dead-letter alert coupling (exact format string plus the
agentos_worker.consumer logger) is kept so alerting still fires; the eval
lane keeps its distinct logger so eval dead-letters stay alert-free.

Update the queue-stream INTERFACE, ARCHITECTURE, and worker README
citations to the methods' new home on StreamConsumer.
